pyproject.toml
[tool.poetry]
name = "baixa-avantio"
version = "0.1.0"
description = ""
authors = ["AngeloGagno <angelogagno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.38"
psycopg2-binary = "^2.9.10"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
